Improve API server setup abstraction in main
